Move config types to pkg/config/types.go instead of pkg/runner/types.go
